tasklist-login

+++ app/Http/Controllers/TasksController.php
@@ -15,15 +15,23 @@ class TasksController extends Controller
      */
     public function index()
     {
-        //タスク一覧を取得
-        $tasks = Task::all();
-        
-        // タスク一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->get();
 
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+
+        // Welcomeビューでそれらを表示
+        return view('welcome', $data);
     }
+
 
     /**
      * Show the form for creating a new resource.
@@ -50,13 +58,13 @@ class TasksController extends Controller
     {   // バリデーション
         $request->validate([
             'status' => 'required|max:10',   
-            'content' => 'required|max:15',
+            'content' => 'required|max:255',
         ]);
         // タスクを作成
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+         $request->user()->tasks()->create([
+            'status' => $request->status,
+            'content' => $request->content,
+        ]);
 
         // トップページへリダイレクトさせる
         return redirect('/');
@@ -70,7 +78,11 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-        $task = Task::findOrFail($id);
+        $task=Task::findOrFail($id);
+        
+        if (\Auth::id() !== $task->user_id) {
+            return redirect('/');
+        }
         
         return view('tasks.show',[
             'task' => $task,
@@ -87,6 +99,10 @@ class TasksController extends Controller
     {
          // idの値でタスクを検索して取得
         $task= Task::findOrFail($id);
+        
+        if (\Auth::id() !== $task->user_id) {
+            return redirect('/');
+        }
 
         // タスク詳細ビューでそれを表示
         return view('tasks.edit', [
@@ -102,13 +118,23 @@ class TasksController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, $id)
-    {   // バリデーション
+    {   
+        // idの値でタスクを検索して取得
+        $task =Task::findOrFail($id);
+        
+        if (\Auth::id() !== $task->user_id) {
+            return redirect('/');
+        }
+        
+        
+        
+        
+        // バリデーション
         $request->validate([
             'status' => 'required|max:10',   
             'content' => 'required|max:15',
         ]);
-        // idの値でタスクを検索して取得
-        $task =Task::findOrFail($id);
+
         // タスクを更新
         $task->status = $request->status;
         $task->content = $request->content;
@@ -126,12 +152,19 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-          // idの値でタスクを検索して取得
+         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
+        
+        if (\Auth::id() !== $task->user_id) {
+            return redirect('/');
+        }
+        
         // メッセージを削除
         $task->delete();
 
         // トップページへリダイレクトさせる
         return redirect('/');
     }
+    
+    
 }
